Update utils, configs, user features, and add new middleware
The changes include a variety of modifications such as correcting a typo in an enum, adding an enum for error messages, new changes to the redis config file, and improving error handling in a helper function. Utils for expressione were also updated, in addition to interfaces for express and user. User models and user related methods in the services were altered for better performance. A new middleware function was created for better error handling. Other modifications include changes in status code responses, the inclusion of await statements, and adjustments in the caching controller and routes.